+++ sdk/docs/pages/typescript/v1-migration.mdx
@@ -61,14 +61,11 @@ Removed `is` and `assert` helpers
 
 # TODO
 
-- Move helpers in inputs.ts
 - TransactionBlock constructor args?
 - Need to export types for blockData
 - Serialize should be updated to allow exporting v2 state
-- Define TypeTag output type rather than input type?
 - Use valibot parser in TypeTagSerializer
 - Should fromB64, toB64, fromHEX, toHEX still be exported from utils?
-- Parse typeArguments from string in input
 
 # @mysten/bcs
