+++ sdk/typescript/src/transactions/utils.ts
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { SuiMoveNormalizedType } from '../client/index.js';
+import { normalizeSuiAddress } from '../utils/sui-types.js';
+import type { CallArg } from './blockData/v2.js';
 
 export function extractMutableReference(
 	normalizedType: SuiMoveNormalizedType,
@@ -36,5 +38,29 @@ export function extractStructTag(
 	if (typeof mutRef === 'object' && 'Struct' in mutRef) {
 		return mutRef;
 	}
+	return undefined;
+}
+
+export function getIdFromCallArg(arg: string | CallArg) {
+	if (typeof arg === 'string') {
+		return normalizeSuiAddress(arg);
+	}
+
+	if (arg.Object) {
+		if (arg.Object.ImmOrOwnedObject) {
+			return normalizeSuiAddress(arg.Object.ImmOrOwnedObject.objectId);
+		}
+
+		if (arg.Object.Receiving) {
+			return normalizeSuiAddress(arg.Object.Receiving.objectId);
+		}
+
+		return normalizeSuiAddress(arg.Object.SharedObject.objectId);
+	}
+
+	if (arg.UnresolvedObject) {
+		return normalizeSuiAddress(arg.UnresolvedObject.objectId);
+	}
+
 	return undefined;
 }
